Inventory Item Rewrite
Completely rewrote inventory item handling so we never have to deal with that absurdity again. Fixed all of the bugs so all previous inventory functions are back in working order now.

Also made cauldron ingredients persistent if you leave house with ingredients in it.
Plus a bunch of bug fixes and QoL updates for the UI handling for it