Pathfinding Algorithm
Algorithm for path calculation is done. Always finds the shortest path avoiding all obstacles in their current place when path is calculated. Can be recalculated on collision to let them walk around other moving objects.
-Collision really should be calculated in advance because it glitches like hell if something moves into the exact tile it's trying to move to while it's already moving.
-Movement speed will also need tweaking.
Currently implemented as a keybind on player just for testing - needs to be added to actual NPCs later.