Build Settings
Disabled ridiculously small screen resolutions - only 16:9 and 16:10 are allowed now just in case.
Added an icon for the .exe and changed the build name so it builds as Hocus Potions.exe now no matter where we build it to.
Also added the default cursor so it'll change as soon as the game starts without having to wait for a script to change it.